dotnet,benchmark,kmgen: [dotnet] improve performance using raw struct instead of opaque pointers
Previously, the Viewer struct is one opaque pointer to the data. Now,
that was changed to "hold" the content and be able to use unsafe to
cast it directly, instead of having one temporary variable.

The performance improves from ~104ms to ~89ms.

The struct uses `Unsafe.AsPointer(this)` to get their own pointer.