kmparser,kmgen: add @key() document tag and @id() struct tag
Before that change, was not possible to specify the id of each struct.
That change makes possible to specify custom Key for the document, and
also IDs for each struct. If no ID is specified, it will be generated
using Blake2b, based on the struct name.

Using `karmmem fmt` will define the `@id()`, based on struct-name.

Signed-off-by: Inkeliz <inkeliz@inkeliz.com>